Support content negotiation on URL download
When downloading data, we can now send an Accept header specifying the
content types we support; the remote end can adapt to this header and
send the right format. For instance, Wikidata’s entity IRIs (e. g.
http://www.wikidata.org/entity/Q42) return JSON by default, but Turtle
when requested with Accept: text/turtle, so this change makes it
possible to use those IRIs for bin/validate’s -d option.

The GET function gains a second parameter, which is optional; omitting
it results in the old behavior. The loadList function signature changes
in an incompatible way, but that function is not exported.