~ fix depth counting in emptyTracker
Arrow functions do not have their own `this`: using `this` in them
always refers to the `this` of the containing scope. To write a method
of an object, and refer to the `depth` of the tracker instead of that of
the surrounding ShExValidator_constructor (which is undefined), ordinary
functions must be used. See also [1].

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/Arrow_functions#Arrow_functions_used_as_methods