Add hideData URL parameter to shex-simple
If the “data” only specifies the SPARQL endpoint, it’s not especially
useful, and if that’s specified via a URL parameter (…&data=Endpoint:
https://example.com/sparql&…), then there’s no need to show it to the
user at all. Add an extra URL parameter, …&hideData&…, to hide the
textarea in that case. (The parameter has no value, its presence or
absence indicates whether the textarea should be hidden or not.)