Hide “Manifest:” text if there are no manifests
If no manifests were loaded (e. g. because &manifest=[] was specified in
the URL), also hide the “Manifest:” text, because it doesn’t make much
sense to display it over an empty list.

To do this, the “Manifest:” text is moved into the div.manifest, which
requires its max-height to be bumped slightly. The visibility of the
surrounding #manifestDrop.droparea isn’t changed, so that drag’n’drop of
manifest JSON files still works (the “Manifest:” text is shown again in
that event).