Improve validate button label when using SPARQL query
When a SPARQL query is used for the shape map, then pressing Ctrl+Enter
might not immediately run validation; it might, instead, first trigger
fetching query results to populate the fixed map, and then a second
Ctrl+Enter will actually validate the fetched entities. In this case, we
change the label of the button to indicate this.

The cleanup code in copyEditMapToFixedMap(), which also removes the
“running” state from the fixed map, is updated to reset the button label
as well, because at this point we’ll always fetch entities. Conversely,
markEditMapDirty() is changed to also change the button label if the
query map is a SPARQL query, and places that use it as an event handler
are switched from the “change” event (only fires once the user has
“committed” a change – in case of a textarea: once focus has left the
element) to “input” (fires on every change, so that the button label
reflects it as soon as possible). And finally, the button label is also
initialized to the SPARQL query version before the first input event.